Capture the datazoom label in hook+component
